+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/AnilistModels.kt
@@ -22,7 +22,7 @@ data class ALManga(
     val format: String,
     val publishing_status: String,
     val start_date_fuzzy: Long,
-    val total_chapters: Int,
+    val total_chapters: Long,
     val average_score: Int,
 ) {
 
@@ -32,7 +32,7 @@ data class ALManga(
         total_chapters = this@ALManga.total_chapters
         cover_url = image_url_lge
         summary = description?.htmlDecode() ?: ""
-        score = average_score.toFloat()
+        score = average_score.toDouble()
         tracking_url = AnilistApi.mangaUrl(remote_id)
         publishing_status = this@ALManga.publishing_status
         publishing_type = format
@@ -55,7 +55,7 @@ data class ALAnime(
     val format: String,
     val publishing_status: String,
     val start_date_fuzzy: Long,
-    val total_episodes: Int,
+    val total_episodes: Long,
     val average_score: Int,
 ) {
 
@@ -65,7 +65,7 @@ data class ALAnime(
         total_episodes = this@ALAnime.total_episodes
         cover_url = image_url_lge
         summary = description?.htmlDecode() ?: ""
-        score = average_score.toFloat()
+        score = average_score.toDouble()
         tracking_url = AnilistApi.animeUrl(remote_id)
         publishing_status = this@ALAnime.publishing_status
         publishing_type = format
@@ -94,10 +94,10 @@ data class ALUserManga(
         remote_id = manga.remote_id
         title = manga.title_user_pref
         status = toTrackStatus()
-        score = score_raw.toFloat()
+        score = score_raw.toDouble()
         started_reading_date = start_date_fuzzy
         finished_reading_date = completed_date_fuzzy
-        last_chapter_read = chapters_read.toFloat()
+        last_chapter_read = chapters_read.toDouble()
         library_id = this@ALUserManga.library_id
         total_chapters = manga.total_chapters
     }
@@ -127,10 +127,10 @@ data class ALUserAnime(
         remote_id = anime.remote_id
         title = anime.title_user_pref
         status = toTrackStatus()
-        score = score_raw.toFloat()
+        score = score_raw.toDouble()
         started_watching_date = start_date_fuzzy
         finished_watching_date = completed_date_fuzzy
-        last_episode_seen = episodes_seen.toFloat()
+        last_episode_seen = episodes_seen.toDouble()
         library_id = this@ALUserAnime.library_id
         total_episodes = anime.total_episodes
     }
