+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mangaupdates/MangaUpdates.kt
@@ -20,11 +20,11 @@ import tachiyomi.domain.track.manga.model.MangaTrack as DomainTrack
 class MangaUpdates(id: Long) : BaseTracker(id, "MangaUpdates"), MangaTracker, DeletableMangaTracker {
 
     companion object {
-        const val READING_LIST = 0
-        const val WISH_LIST = 1
-        const val COMPLETE_LIST = 2
-        const val UNFINISHED_LIST = 3
-        const val ON_HOLD_LIST = 4
+        const val READING_LIST = 0L
+        const val WISH_LIST = 1L
+        const val COMPLETE_LIST = 2L
+        const val UNFINISHED_LIST = 3L
+        const val ON_HOLD_LIST = 4L
 
         private val SCORE_LIST = (0..10)
             .flatMap { decimal ->
@@ -47,11 +47,11 @@ class MangaUpdates(id: Long) : BaseTracker(id, "MangaUpdates"), MangaTracker, De
 
     override fun getLogoColor(): Int = Color.rgb(146, 160, 173)
 
-    override fun getStatusListManga(): List<Int> {
+    override fun getStatusListManga(): List<Long> {
         return listOf(READING_LIST, COMPLETE_LIST, ON_HOLD_LIST, UNFINISHED_LIST, WISH_LIST)
     }
 
-    override fun getStatus(status: Int): StringResource? = when (status) {
+    override fun getStatus(status: Long): StringResource? = when (status) {
         READING_LIST -> MR.strings.reading_list
         WISH_LIST -> MR.strings.wish_list
         COMPLETE_LIST -> MR.strings.complete_list
@@ -60,15 +60,15 @@ class MangaUpdates(id: Long) : BaseTracker(id, "MangaUpdates"), MangaTracker, De
         else -> null
     }
 
-    override fun getReadingStatus(): Int = READING_LIST
+    override fun getReadingStatus(): Long = READING_LIST
 
-    override fun getRereadingStatus(): Int = -1
+    override fun getRereadingStatus(): Long = -1
 
-    override fun getCompletionStatus(): Int = COMPLETE_LIST
+    override fun getCompletionStatus(): Long = COMPLETE_LIST
 
     override fun getScoreList(): ImmutableList<String> = SCORE_LIST
 
-    override fun indexToScore(index: Int): Float = if (index == 0) 0f else SCORE_LIST[index].toFloat()
+    override fun indexToScore(index: Int): Double = if (index == 0) 0.0 else SCORE_LIST[index].toDouble()
 
     override fun displayScore(track: DomainTrack): String = track.score.toString()
 
@@ -89,7 +89,7 @@ class MangaUpdates(id: Long) : BaseTracker(id, "MangaUpdates"), MangaTracker, De
             val (series, rating) = api.getSeriesListItem(track)
             track.copyFrom(series, rating)
         } catch (e: Exception) {
-            track.score = 0f
+            track.score = 0.0
             api.addSeriesToList(track, hasReadChapters)
             track
         }
@@ -109,7 +109,7 @@ class MangaUpdates(id: Long) : BaseTracker(id, "MangaUpdates"), MangaTracker, De
 
     private fun MangaTrack.copyFrom(item: ListItem, rating: Rating?): MangaTrack = apply {
         item.copyTo(this)
-        score = rating?.rating ?: 0f
+        score = rating?.rating ?: 0.0
     }
 
     override suspend fun login(username: String, password: String) {
